refactor: move test code from ChildProcessError to GetHelpfulChildProcessError
